refactor: big refactor
