--- frontend/src/routes/dashboard.lazy.tsx
@@ -1,10 +1,0 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
-import { withAuthRedirect } from "~/components/helper/auth-redirect";
-
-export const Route = createLazyFileRoute("/dashboard")({
-  component: withAuthRedirect(Dashboard)
-});
-
-function Dashboard() {
-  return <div></div>;
-}
